split Grid out from Bench
